fix code

+++ src/context_free_lang/canonical_lr_grammar.cpp
@@ -115,7 +115,7 @@ namespace cyy::computation {
         for (const auto &lookahead : lookahead_set) {
           // conflict
           if (action_table.count({state, lookahead}) != 0) {
-            std::cerr << "conflict with follow_terminal" << lookahead
+            std::cerr << "conflict with follow_terminal" << static_cast<int>(lookahead)
                       << std::endl;
             throw cyy::computation::exception::no_canonical_LR_grammar("");
           }
